+++ app/src/plugins/windows/mod.rs
@@ -5,3 +5,4 @@ pub mod udp_listen;
 pub mod autoruns;
 pub mod yara;
 pub mod domain_users;
+pub mod domain_computers;