+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [v0.0.3] - TBD
 ### Added
-- Windows plugins (domain_users)
+- Windows plugins (domain_users, domain_computers)
 
 ### Changed
 - Documentation updated